[Chore] 이름 변경 Option -> Filter (#16)
